Moved service logic out of API
